bot/requirements.txt
urllib3==1.26.15
python-telegram-bot==13.7
psycopg2==2.9.9
python-dotenv==1.0.1
paramiko==3.4.0
requests==2.31.0
